+++ src/Period.php
@@ -723,14 +723,14 @@ final class Period
      *
      * @param \League\Period\Period $period
      *
-     * @throws \InvalidArgumentException
+     * @throws \LogicException
      *
      * @return \League\Period\Period
      */
     public function intersect(Period $period)
     {
         if (! $this->overlaps($period)) {
-            throw new InvalidArgumentException('Both Period objects should overlaps');
+            throw new LogicException('Both Period objects should overlaps');
         }
 
         $start = $this->start;
@@ -751,14 +751,14 @@ final class Period
      *
      * @param \League\Period\Period $period
      *
-     * @throws \InvalidArgumentException
+     * @throws \LogicException
      *
      * @return \League\Period\Period
      */
     public function gap(Period $period)
     {
         if ($this->overlaps($period)) {
-            throw new InvalidArgumentException('Both Period objects should not overlaps');
+            throw new LogicException('Both Period objects should not overlaps');
         }
 
         $start = $this->start;